feat(git): Use Catppuccin as delta theme
This commit modifies the update_themes script to include updates for
delta themes. It clones the catppuccin/delta repository and moves
the themes to ~/.config/delta/themes. It also updates the gitconfig
file to use the Catppuccin theme for delta. The delta features
setting is updated to use Catppuccin Macchiato.