feat: add GitHub dashboard configuration
Added a new configuration file for the GitHub dashboard to streamline the
management of pull requests and issues. This setup includes sections for
personal PRs, review requests, and issues, along with default settings for
preview, layout, and keybindings. The configuration aims to enhance
productivity by providing a customized and organized view of GitHub
activities directly within the dashboard.

https://github.com/dlvhdr/gh-dash